requirements.txt
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.6.0
et-xmlfile==1.0.1
html5lib==1.1
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
Pillow==8.1.0
pycparser==2.20
Pyphen==0.10.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
tinycss2==1.1.0
WeasyPrint==52.2
webencodings==0.5.1
xlrd==2.0.1
